requirements.txt
transformers
pillow
streamlit
torch
sentencepiece
google-cloud-translate==2.0.1
